refactor(cli): modularize command router
