+++ .changeset/modular-cli-router.md
@@ -1,0 +1,5 @@
+---
+"skillset": patch
+---
+
+Modularize the CLI router behind command-owned request and execution seams without changing the public command contract.
